Support `ConnectionPool` (#4)
Add support for `ConnectionPool`, which allows Redis connections to be shared efficiently among threads.  Because this gem hadn't been updated in a long time, it required some maintenance before adding the new feature:

- Update `.gemspec` and `require` statements based on the current state of dependent gems.
- Add the `simplecov` gem and require 100% coverage.
- Run `transpec` to update to RSpec 3.x syntax.

To test the new functionality, the existing spec was extracted into a shared example, which is then invoked with a simple `Redis` instance (the old behavior) and with a `ConnectionPool`.